refactor(api): move fee routes to core
