+++ crates/api/src/lib.rs
@@ -42,10 +42,10 @@ use utoipa::OpenApi;
         routes::aggregated::block_transactions_aggregated,
         routes::core::sequencer_distribution,
         routes::core::sequencer_blocks,
-        routes::aggregated::l2_fees,
-        routes::aggregated::batch_fees,
-        routes::aggregated::l2_fee_components,
-        routes::aggregated::batch_fee_components,
+        routes::core::l2_fees,
+        routes::core::batch_fees,
+        routes::core::l2_fee_components,
+        routes::core::batch_fee_components,
         routes::aggregated::batch_fee_components_aggregated,
         routes::aggregated::dashboard_data,
         routes::core::l1_data_cost,
