+++ crates/api/src/routes/core.rs
@@ -10,18 +10,20 @@ use crate::{
 };
 use alloy_primitives::Address;
 use api_types::{
-    ActiveGatewaysResponse, AvgBlobsPerBatchResponse, BatchPostingTimesResponse, BlockProfitItem,
-    BlockProfitsResponse, ErrorResponse, EthPriceResponse, L1BlockTimesResponse,
-    L1DataCostResponse, L1HeadBlockResponse, L1HeadResponse, L2HeadBlockResponse, L2HeadResponse,
-    ProveCostResponse, ProveTimesResponse, SequencerBlocksItem, SequencerBlocksResponse,
-    SequencerDistributionItem, SequencerDistributionResponse, VerifyTimesResponse,
+    ActiveGatewaysResponse, AvgBlobsPerBatchResponse, BatchFeeComponentRow,
+    BatchFeeComponentsResponse, BatchPostingTimesResponse, BlockProfitItem, BlockProfitsResponse,
+    ErrorResponse, EthPriceResponse, FeeComponentsResponse, L1BlockTimesResponse,
+    L1DataCostResponse, L1HeadBlockResponse, L1HeadResponse, L2FeesResponse, L2HeadBlockResponse,
+    L2HeadResponse, ProveCostResponse, ProveTimesResponse, SequencerBlocksItem,
+    SequencerBlocksResponse, SequencerDistributionItem, SequencerDistributionResponse,
+    SequencerFeeRow, VerifyTimesResponse,
 };
 use axum::{
     Json,
     extract::{Query, State},
     http::StatusCode,
 };
-use clickhouse_lib::{AddressBytes, L1DataCostRow, ProveCostRow};
+use clickhouse_lib::{AddressBytes, BlockFeeComponentRow, L1DataCostRow, ProveCostRow};
 use hex::encode;
 use primitives::WEI_PER_GWEI;
 
@@ -655,4 +657,277 @@ pub async fn eth_price(
             ))
         }
     }
+}
+
+#[utoipa::path(
+    get,
+    path = "/l2-fees",
+    params(
+        RangeQuery
+    ),
+    responses(
+        (status = 200, description = "Priority and base fees", body = L2FeesResponse),
+        (status = 500, description = "Database error", body = ErrorResponse)
+    ),
+    tag = "taikoscope"
+)]
+/// Get L2 fee breakdown including priority fees, base fees, and L1 data costs by sequencer
+pub async fn l2_fees(
+    Query(params): Query<RangeQuery>,
+    State(state): State<ApiState>,
+) -> Result<Json<L2FeesResponse>, ErrorResponse> {
+    validate_time_range(&params.time_range)?;
+
+    let has_time_range = has_time_range_params(&params.time_range);
+    validate_range_exclusivity(has_time_range, false)?;
+
+    let time_range = resolve_time_range_enum(&params.range, &params.time_range);
+    let address = if let Some(addr) = params.address.as_ref() {
+        match addr.parse::<Address>() {
+            Ok(a) => Some(AddressBytes::from(a)),
+            Err(e) => {
+                tracing::warn!(error = %e, "Failed to parse address");
+                return Err(ErrorResponse::new(
+                    "invalid-params",
+                    "Bad Request",
+                    StatusCode::BAD_REQUEST,
+                    e.to_string(),
+                ));
+            }
+        }
+    } else {
+        None
+    };
+
+    let (priority_fee, base_fee, l1_data_cost, prove_cost, rows) = tokio::try_join!(
+        state.client.get_l2_priority_fee(address, time_range),
+        state.client.get_l2_base_fee(address, time_range),
+        state.client.get_l1_total_data_cost(address, time_range),
+        state.client.get_total_prove_cost(address, time_range),
+        state.client.get_l2_fees_by_sequencer(time_range)
+    )
+    .map_err(|e| {
+        tracing::error!(error = %e, "Failed to get L2 fees");
+        ErrorResponse::database_error()
+    })?;
+
+    let sequencers: Vec<SequencerFeeRow> = rows
+        .into_iter()
+        .filter(|r| if let Some(target) = address { r.sequencer == target } else { true })
+        .map(|r| SequencerFeeRow {
+            address: format!("0x{}", encode(r.sequencer)),
+            priority_fee: r.priority_fee / WEI_PER_GWEI,
+            base_fee: r.base_fee / WEI_PER_GWEI,
+            l1_data_cost: r.l1_data_cost.map(|v| v / WEI_PER_GWEI),
+            prove_cost: r.prove_cost.map(|v| v / WEI_PER_GWEI),
+        })
+        .collect();
+
+    let priority_fee = priority_fee.map(|v| v / WEI_PER_GWEI);
+    let base_fee = base_fee.map(|v| v / WEI_PER_GWEI);
+    let l1_data_cost = l1_data_cost.map(|v| v / WEI_PER_GWEI);
+    let prove_cost = prove_cost.map(|v| v / WEI_PER_GWEI);
+
+    tracing::info!(count = sequencers.len(), "Returning L2 fees and breakdown");
+    Ok(Json(L2FeesResponse { priority_fee, base_fee, l1_data_cost, prove_cost, sequencers }))
+}
+
+#[utoipa::path(
+    get,
+    path = "/batch-fees",
+    params(
+        RangeQuery
+    ),
+    responses(
+        (status = 200, description = "Priority and base fees per batch", body = L2FeesResponse),
+        (status = 500, description = "Database error", body = ErrorResponse)
+    ),
+    tag = "taikoscope"
+)]
+/// Get batch fee breakdown including priority fees, base fees, and L1 data costs by proposer
+pub async fn batch_fees(
+    Query(params): Query<RangeQuery>,
+    State(state): State<ApiState>,
+) -> Result<Json<L2FeesResponse>, ErrorResponse> {
+    validate_time_range(&params.time_range)?;
+
+    let has_time_range = has_time_range_params(&params.time_range);
+    validate_range_exclusivity(has_time_range, false)?;
+
+    let time_range = resolve_time_range_enum(&params.range, &params.time_range);
+    let address = if let Some(addr) = params.address.as_ref() {
+        match addr.parse::<Address>() {
+            Ok(a) => Some(AddressBytes::from(a)),
+            Err(e) => {
+                tracing::warn!(error = %e, "Failed to parse address");
+                return Err(ErrorResponse::new(
+                    "invalid-params",
+                    "Bad Request",
+                    StatusCode::BAD_REQUEST,
+                    e.to_string(),
+                ));
+            }
+        }
+    } else {
+        None
+    };
+
+    let (priority_fee, base_fee, l1_data_cost, rows) = tokio::try_join!(
+        state.client.get_batch_priority_fee(address, time_range),
+        state.client.get_batch_base_fee(address, time_range),
+        state.client.get_batch_total_data_cost(address, time_range),
+        state.client.get_batch_fees_by_proposer(time_range)
+    )
+    .map_err(|e| {
+        tracing::error!(error = %e, "Failed to get batch fees");
+        ErrorResponse::database_error()
+    })?;
+
+    let sequencers: Vec<SequencerFeeRow> = rows
+        .into_iter()
+        .filter(|r| if let Some(target) = address { r.sequencer == target } else { true })
+        .map(|r| SequencerFeeRow {
+            address: format!("0x{}", encode(r.sequencer)),
+            priority_fee: r.priority_fee / WEI_PER_GWEI,
+            base_fee: r.base_fee / WEI_PER_GWEI,
+            l1_data_cost: r.l1_data_cost.map(|v| v / WEI_PER_GWEI),
+            prove_cost: r.prove_cost.map(|v| v / WEI_PER_GWEI),
+        })
+        .collect();
+
+    let priority_fee = priority_fee.map(|v| v / WEI_PER_GWEI);
+    let base_fee = base_fee.map(|v| v / WEI_PER_GWEI);
+    let l1_data_cost = l1_data_cost.map(|v| v / WEI_PER_GWEI);
+
+    tracing::info!(count = sequencers.len(), "Returning batch fees and breakdown");
+    Ok(Json(L2FeesResponse { priority_fee, base_fee, l1_data_cost, prove_cost: None, sequencers }))
+}
+
+#[utoipa::path(
+    get,
+    path = "/l2-fee-components",
+    params(
+        RangeQuery
+    ),
+    responses(
+        (status = 200, description = "Fee components per block", body = FeeComponentsResponse),
+        (status = 500, description = "Database error", body = ErrorResponse)
+    ),
+    tag = "taikoscope"
+)]
+/// Get detailed fee components per block showing priority fee, base fee, and L1 data cost
+pub async fn l2_fee_components(
+    Query(params): Query<RangeQuery>,
+    State(state): State<ApiState>,
+) -> Result<Json<FeeComponentsResponse>, ErrorResponse> {
+    validate_time_range(&params.time_range)?;
+
+    let has_time_range = has_time_range_params(&params.time_range);
+    validate_range_exclusivity(has_time_range, false)?;
+
+    let time_range = resolve_time_range_enum(&params.range, &params.time_range);
+    let address = if let Some(addr) = params.address.as_ref() {
+        match addr.parse::<Address>() {
+            Ok(a) => Some(AddressBytes::from(a)),
+            Err(e) => {
+                tracing::warn!(error = %e, "Failed to parse address");
+                return Err(ErrorResponse::new(
+                    "invalid-params",
+                    "Bad Request",
+                    StatusCode::BAD_REQUEST,
+                    e.to_string(),
+                ));
+            }
+        }
+    } else {
+        None
+    };
+
+    let blocks = state.client.get_l2_fee_components(address, time_range).await.map_err(|e| {
+        tracing::error!(error = %e, "Failed to get fee components");
+        ErrorResponse::database_error()
+    })?;
+
+    let blocks: Vec<BlockFeeComponentRow> = blocks
+        .into_iter()
+        .map(|r| BlockFeeComponentRow {
+            l2_block_number: r.l2_block_number,
+            priority_fee: r.priority_fee / WEI_PER_GWEI,
+            base_fee: r.base_fee / WEI_PER_GWEI,
+            l1_data_cost: r.l1_data_cost.map(|v| v / WEI_PER_GWEI),
+        })
+        .collect();
+
+    Ok(Json(FeeComponentsResponse { blocks }))
+}
+
+#[utoipa::path(
+    get,
+    path = "/batch-fee-components",
+    params(
+        RangeQuery
+    ),
+    responses(
+        (status = 200, description = "Fee components per batch", body = BatchFeeComponentsResponse),
+        (status = 500, description = "Database error", body = ErrorResponse)
+    ),
+    tag = "taikoscope"
+)]
+/// Get detailed fee components per batch showing priority fee, base fee, and L1 data cost
+pub async fn batch_fee_components(
+    Query(params): Query<RangeQuery>,
+    State(state): State<ApiState>,
+) -> Result<Json<BatchFeeComponentsResponse>, ErrorResponse> {
+    validate_time_range(&params.time_range)?;
+
+    let has_time_range = has_time_range_params(&params.time_range);
+    validate_range_exclusivity(has_time_range, false)?;
+
+    let time_range = resolve_time_range_enum(&params.range, &params.time_range);
+    let address = if let Some(addr) = params.address.as_ref() {
+        match addr.parse::<Address>() {
+            Ok(a) => Some(AddressBytes::from(a)),
+            Err(e) => {
+                tracing::warn!(error = %e, "Failed to parse address");
+                return Err(ErrorResponse::new(
+                    "invalid-params",
+                    "Bad Request",
+                    StatusCode::BAD_REQUEST,
+                    e.to_string(),
+                ));
+            }
+        }
+    } else {
+        None
+    };
+
+    let rows = state.client.get_batch_fee_components(address, time_range).await.map_err(|e| {
+        tracing::error!(error = %e, "Failed to get batch fee components");
+        ErrorResponse::database_error()
+    })?;
+
+    let prove_total =
+        state.client.get_total_prove_cost(address, time_range).await.map_err(|e| {
+            tracing::error!(error = %e, "Failed to get prove cost");
+            ErrorResponse::database_error()
+        })?;
+    let count = rows.len() as u128;
+    let amortized_prove =
+        if count > 0 { prove_total.map(|c| (c / count) / WEI_PER_GWEI) } else { None };
+
+    let batches: Vec<BatchFeeComponentRow> = rows
+        .into_iter()
+        .map(|r| BatchFeeComponentRow {
+            batch_id: r.batch_id,
+            l1_block_number: r.l1_block_number,
+            l1_tx_hash: format!("0x{}", encode(r.l1_tx_hash)),
+            sequencer: format!("0x{}", encode(r.sequencer)),
+            priority_fee: r.priority_fee / WEI_PER_GWEI,
+            base_fee: r.base_fee / WEI_PER_GWEI,
+            l1_data_cost: r.l1_data_cost.map(|v| v / WEI_PER_GWEI),
+            amortized_prove_cost: amortized_prove,
+        })
+        .collect();
+
+    Ok(Json(BatchFeeComponentsResponse { batches }))
 }
